Fix build with nmake.

+++ gzwrite.c
@@ -194,7 +194,7 @@ static size_t gz_write(gz_statep state, void const *buf, size_t len)
                               state->in);
             copy = state->size - have;
             if (copy > len)
-                copy = len;
+                copy = (unsigned)len;
             memcpy(state->in + have, buf, copy);
             state->strm.avail_in += copy;
             state->x.pos += copy;
@@ -213,7 +213,7 @@ static size_t gz_write(gz_statep state, void const *buf, size_t len)
         do {
             unsigned n = -1;
             if (n > len)
-                n = len;
+                n = (unsigned)len;
             state->strm.avail_in = n;
             state->x.pos += n;
             if (gz_comp(state, Z_NO_FLUSH) == -1)
@@ -338,7 +338,7 @@ int ZEXPORT gzputs(gzFile file, const char *str) {
 
     /* write string */
     len = strlen(str);
-    ret = gz_write(state, str, len);
+    ret = (int)gz_write(state, str, len);
     return ret == 0 && len != 0 ? -1 : ret;
 }
 
@@ -380,7 +380,7 @@ int ZEXPORTVA gzvprintf(gzFile file, const char *format, va_list va) {
     len = vsnprintf(next, state->size, format, va);
 
     /* check that printf() results fit in buffer */
-    if (len == 0 || len >= state->size || next[state->size - 1] != 0)
+    if (len == 0 || (unsigned)len >= state->size || next[state->size - 1] != 0)
         return 0;
 
     /* update buffer and position, compress first half if past that */
